+++ src/ml_tooling/plots/feature_importance.py
@@ -1,22 +1,26 @@
+"""Low-level implementation of plotting feature importance from an estimator"""
+
 from typing import Union
 
+import numpy as np
 import pandas as pd
 from matplotlib.axes import Axes
+from sklearn.base import BaseEstimator
 
-from sklearn.inspection import permutation_importance
 from ml_tooling.plots.utils import _plot_barh
-from ml_tooling.utils import DataType, Estimator
-from sklearn.base import is_classifier
+from ml_tooling.utils import (
+    DataType,
+    Estimator,
+    is_pipeline,
+    _get_labels_from_pipeline,
+    VizError,
+)
 
 
 def plot_feature_importance(
     estimator: Estimator,
     x: pd.DataFrame,
     y: DataType,
-    scoring="default",
-    n_repeats: int = 5,
-    n_jobs: int = -1,
-    random_state: int = None,
     ax: Axes = None,
     bottom_n: Union[int, float] = None,
     top_n: Union[int, float] = None,
@@ -24,28 +28,79 @@ def plot_feature_importance(
     title: str = "",
     **kwargs,
 ) -> Axes:
-    if scoring == "default":
-        scoring = "accuracy" if is_classifier(estimator) else "r2"
+    """
+    Plot either the estimator coefficients or the estimator
+    feature importances depending on what is provided by the estimator.
 
-    importances = permutation_importance(
-        estimator,
-        x,
-        y,
-        scoring=scoring,
-        n_repeats=n_repeats,
-        n_jobs=n_jobs,
-        random_state=random_state,
-    )
+    see also :func:ml_tooling.plot.plot_permutation_importance for an
+    unbiased version of feature importance using permutation importance
 
-    feature_importance = importances.importances_mean
-    labels = x.columns
-    plot_title = title if title else f"Feature Importances ({scoring.title()})"
+
+    Parameters
+    ----------
+    estimator: Estimator
+        Estimator to use to calculate permuted feature importance
+
+    x: DataType
+        Features to calculate permuted feature importance for
+
+    y: DataType
+        Target to use in scoring
+
+    ax: Axes
+        Matplotlib axes to draw the graph on. Creates a new one by default
+
+    bottom_n: int
+        Plot only bottom n features
+
+    top_n: int
+        Plot only top n features
+
+    add_label: bool
+        Whether or not to plot text labels for the bars
+
+    title: str
+        Title to add to the plot
+
+    kwargs: dict
+        Any kwargs are passed to matplotlib
+
+    Returns
+    -------
+    plt.Axes
+
+    """
+
+    trained_estimator: BaseEstimator = estimator.steps[-1][1] if is_pipeline(
+        estimator
+    ) else estimator
+
+    if hasattr(trained_estimator, "coef_"):
+        feature_importances: np.ndarray = getattr(trained_estimator, "coef_").squeeze()
+        x_label = "Coefficients"
+
+    elif hasattr(trained_estimator, "feature_importances_"):
+        feature_importances: np.ndarray = getattr(
+            trained_estimator, "feature_importances_"
+        )
+        x_label = "Feature Importances"
+
+    else:
+        raise VizError(
+            "Estimator must have one of coef_ or feature_importances_."
+            f"{estimator} has neither. Make sure that it has been fitted"
+        )
+
+    labels = _get_labels_from_pipeline(estimator, x)
+
+    plot_title = title if title else "Feature Importances"
+
     ax = _plot_barh(
-        feature_importance,
+        feature_importances.squeeze(),
         labels,
         add_label=add_label,
         title=plot_title,
-        x_label=f"Permuted Feature Importance ({scoring.title()}) Relative to Baseline",
+        x_label=x_label,
         y_label="Feature Labels",
         ax=ax,
         top_n=top_n,