+++ src/ml_tooling/search/__init__.py
@@ -1,4 +1,13 @@
 from .gridsearch import GridSearch
 from .randomsearch import RandomSearch
+from .bayes import BayesSearch
+from skopt.space import Real, Categorical, Integer
 
-__all__ = ["GridSearch", "RandomSearch"]
+__all__ = [
+    "GridSearch",
+    "RandomSearch",
+    "BayesSearch",
+    "Real",
+    "Categorical",
+    "Integer",
+]